LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T3-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S12.14_A16_T3

PASS Expected parsing failure

TEST COMPLETE
